Add files via upload
VGG16 is a convolutional neural network architecture that was introduced by the Visual Geometry Group (VGG) at the University of Oxford. It was one of the contestants in the ImageNet Large Scale Visual Recognition Challenge (ILSVRC) in 2014. The "16" in VGG16 refers to the number of weight layers in the network.